基本完成

+++ src/main/java/com/liuhuang/voteprogram/dto/AnonymousVoteCountResponseDTO.java
@@ -6,10 +6,10 @@ import lombok.Data;
 @Data
 public class AnonymousVoteCountResponseDTO {
     private Long count;
-    private Long optionId;
+    private int optionId;
     private String optionText;
 
-    public AnonymousVoteCountResponseDTO(Long count, Long optionId, String optionText) {
+    public AnonymousVoteCountResponseDTO(Long count, int optionId, String optionText) {
         this.count = count;
         this.optionId = optionId;
         this.optionText = optionText;
